r4-p2-dapr-actors-stream-injector/financial-advisor-agent/pyproject.toml
[project]
name = "financial-advisor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.3",
    "cloudevents>=1.12.0",
    "dapr>=1.15.0",
    "dapr-ext-fastapi>=1.15.0",
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.2.8",
    "redis>=6.4.0",
]

[dependency-groups]
dev = []
